% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dematel_spectral.R
\name{print.dematel_spectral}
\alias{print.dematel_spectral}
\title{Print method for dematel_spectral objects}
\usage{
\method{print}{dematel_spectral}(x, ...)
}
\description{
Print method for dematel_spectral objects
}
